I am adding in the backend server to our Angular application.
index.ts: Barebones server setup, definition of variables (expense date, amount, etc.), and simple CRUD operations.

expenses.json: Sample expense data to be used to create the table on app initialization.

.gitignore: Adding the /src/server/node_modules directory into this file.

My understanding is that we have to run 'npm install' both in the root directory and in the /src/server directory.

My next step will be using the 'concurrently' tool to set it up so that using npm start will both serve the Angular frontend and start up the ExpressJS backend at the same time. Perhaps this can also be used for the npm install process...